Remove default constructor. Add constructor with parameteres to Particle.

+++ src/particle.cpp
@@ -1,6 +1,7 @@
 #include "particle.h"
+#include <iostream>
 
-Particle::Particle(const ofPoint& position, const ofPoint& velocity, int lifeTime) {
+Particle::Particle(const ofPoint& position, const ofPoint& velocity, float lifeTime) {
     this->position = position;
     this->velocity = velocity;
